gp11/gp11.h
	* gp11/gp11-attributes.c:
	* gp11/gp11-misc.c:
	* gp11/gp11-object.c:
	* gp11/gp11-private.h:
	* gp11/gp11-session.c:
	* gp11/gp11-slot.c:
	* gp11/gp11.h
	* gp11/tests/unit-test-gp11-object.c:
	* gp11/tests/unit-test-gp11-slot.c: 
	* tool/gkr-tool-import.c: Use gulong for all CK_ULONG derived
	types in PKCS#11. Fixes bug #547830

svn path=/trunk/; revision=1248

+++ gp11/gp11-misc.c
@@ -199,7 +199,7 @@ gp11_message_from_rv (CK_RV rv)
 		return _("The signature request was rejected by the user");
 		
 	default:
-		g_message ("unknown error: %u", (guint)rv);
+		g_message ("unknown error: %lu", (gulong)rv);
 		return _("Unknown error");
 	}
 }
